fix: database population bugs

+++ database_population.py
@@ -187,7 +187,7 @@ def populate_trades(connection):
         time_executed = row['Filing Date']
         time_executed = convert_dmy_to_ymd(time_executed)
         time_offered = convert_dmy_to_ymd(time_offered)
-        trade_json = json.dumps({ "type": type, "symbol": symbol, "quantity": int(quantity * 100), "price_per_item": int(price_per_item * 100), "time_offered": time_offered})
+        trade_json = json.dumps({ "type": type, "symbol": symbol, "quantity": int(float(quantity) * 100), "price_per_item": int(float(price_per_item)* 100), "time_offered": time_offered})
         if username not in data_trades:
             data_trades[username] = {}
         
@@ -203,9 +203,9 @@ def populate_portfolio(connection):
             date = date.decode('utf-8')
             trade = json.loads(trade.decode('utf-8'))
             symbol = trade['symbol']
-            quantity = float(trade['quantity'])
+            quantity = int(trade['quantity'])
             type = trade['type']
-            price_per_item = float(trade['price_per_item'])
+            price_per_item = int(trade['price_per_item'])
             if symbol not in user_stocks:
                 user_stocks[symbol] = tuple([0,0])
             if type == "P":
@@ -247,8 +247,8 @@ def populate_popularity_to_instrument(connection):
 
             #get the trade information
             trade = json.loads(trade.decode('utf-8'))
-            symbol, quantity = trade['symbol'], float(trade['quantity'])
-            price = float(trade['price_per_item'])
+            symbol, quantity = trade['symbol'], int(trade['quantity'])
+            price = int(trade['price_per_item'])
             cost_of_trade = quantity * price / 100
 
             #calculate the timestamp score based on the number of years since 2020 plus the cost of the trade (/ (3600*24*365) converts to years)
